requirements.txt
configparser==3.5.0
Django==1.10.5
djangorestframework==3.5.3
mysqlclient==1.3.9
